デバッグモード時に画像/txtを入れたディレクトリを出力するよう変更

+++ src/python/converter/core/image_converter.py
@@ -87,8 +87,13 @@ def convert_image_parallel(cfg: AppConfig, img_info: list[int, str, str]) -> Non
     # 末尾に独自データを追記
     append_footer_data(temppng_path, tempbin_path)
 
-    # 変換し終わったpngファイルを削除
-    temppng_path.unlink()
+    if (cfg.debug_mode):
+        # デバッグ用に中間pngファイルをdebug_dirにコピー
+        debug_png_path = cfg.debug_dir / f'img{p_index}.png'
+        temppng_path.replace(debug_png_path)
+    else:
+        # 変換し終わったpngファイルを削除
+        temppng_path.unlink()
 
     # 競合するファイルがあれば削除
     if output_path.exists():
